When the editor level is created, a blank default texture layer is automatically created
- Reduce memory consumption of a texture layer by 50% using uint16_t instead of int; 16 bit types are enough to lower hard limits regarding tileset dimensions and dep-table entry counts

+++ Marble Engine/include/level.h
@@ -66,15 +66,15 @@ union marble_levelasset_layerobj {
 	 *     y := { 0, height - 1 }
 	 */
 	struct marble_levelasset_texobj {
-		int m_depindex; /* index into the dep-table */
+		uint16_t m_depindex; /* index into the dep-table */
 		/*
 		 * index into the dependency
 		 * 
 		 * The index is represented by a format
 		 * similar to the one described by **i**.
 		 */
-		int m_x;
-		int m_y;
+		uint16_t m_x;
+		uint16_t m_y;
 	} _tex;
 };
